+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## 7.1.3
 
 * `Versions`:
-  * `Serialization`: `1.5.0` -> `1.5.1`
-  * `MicroUtils`: `0.18.1` -> `0.18.4`
+    * `Serialization`: `1.5.0` -> `1.5.1`
+    * `MicroUtils`: `0.18.1` -> `0.18.4`
+* `BehaviourBuilder`:
+    * Now it is possible to use `waitMediaContent`/`waitMediaContentMessage`/`onMediaContent`
 
 ## 7.1.2